feat(security): Replace security-proxy-setup for adding users (#2808)
Replaces security-proxy-setup for adding users to the Kong proxy.

This new utility changes the JWT authentication mode from HS256 to
RS256, thus avoiding exposure of the HMAC secret on the server side.
This new utility also puts the client_id and client_secret for
OAuth2 authentication in the hands of the user.
Although in the new scheme the user is responsible for generating
tokens via either the client_credentials OAuth2 flow,
or client-side JWT signing, utility methods are provided
for that function that do not depend upon the server.

The new utility eliminates the requirement from security-proxy-setup
that a valid security secret store token is available.

Fixes #2477
Fixes #2549
Fixes #2568

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>