fix(snap): restrict consul to localhost
This commit updates the consul startup script to
no longer specify a client IPv4 address, which
causes consul to default to localhost, effectively
preventing consul access from another system. This
change also adds support for the a new config option:

env.security-proxy.add-proxy-route

...which can be used to add additional services to
the API Gateway. If add-proxy-route is not configured
prior to snap install, it will be initialized at install
time to:

consul.http://localhost:8500

This causes a proxy route to be setup for Consul along
with all of the default services.

Signed-off-by: Tony Espy <espy@canonical.com>