feat(snap): add go-based configure and install hooks
This commit

- adds go-based configure and install hooks to the snap,
including support for service-specific configuration overrides via
service.env files read by services on startup.

- adds the ability to configure the API Gateway using snap configuration

- adds logic to the setup-consul-acl script
(which launches security-bootstrapper) to source the
associated .env file which is used to set env var used
to configure additional consul ACLS.

- updates the new options to use the service
name "security-secret-store" vs. "security-secrestore".

- updates README configuration text

- updates the snap to use a wrapper script
for the security-bootstrapper-redis service (one-shot)
so that the hard-coded path in the generated redis.conf
file can be updated to use a symlink instead of a hard-
coded revision in the path to the ACL file.

- creates an ephemeral top-level
VERSION file in the edgex-go build dir so
that the services log the correct verison
on startup.

For more information refer to the snap/README.md file.

Signed-off-by: Tony Espy <espy@canonical.com>

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>